+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Kotlin_Bromine_Arya/Tele/SubSystems/Shoulder.kt
@@ -4,7 +4,7 @@ import com.qualcomm.robotcore.hardware.DcMotor
 import com.qualcomm.robotcore.hardware.DcMotorEx
 import com.qualcomm.robotcore.hardware.DcMotorSimple
 import com.qualcomm.robotcore.hardware.HardwareMap
-import org.firstinspires.ftc.teamcode.Kotlin_Bromine_Arya.Auto.PID_Components.PIDController
+import org.firstinspires.ftc.teamcode.Kotlin_Bromine_Arya.Util.PIDController
 
 
 open class Shoulder(hwMap: HardwareMap) {
